test_task_111/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
